Debug download paper

+++ download_paper.sh
@@ -2,7 +2,8 @@
 
 export START=$1
 export END=$2
-export DIRECTORY=$3
+export SLEEP=30
+export DIRECTORY="paper"
 
 # create directory if does not exist
 if [ ! -d $DIRECTORY ]; then
@@ -30,10 +31,22 @@ while (( $START < $END ))
 do
     get_nber_id
     URL="https://www.nber.org/system/files/working_papers/w${NBER_ID}/w${NBER_ID}.pdf"
-    echo $URL
-    wget $URL -O "${DIRECTORY}/${NBER_ID}.pdf"
-    pdftotext -layout "${DIRECTORY}/${NBER_ID}.pdf" "${DIRECTORY}/${NBER_ID}.txt"
-    rm -rf "${DIRECTORY}/${NBER_ID}.pdf"
+    if [ -e "${DIRECTORY}/${NBER_ID}.txt" ]
+    then
+        echo "${DIRECTORY}/${NBER_ID}.txt already exists."
+    else
+        wget $URL -O "${DIRECTORY}/${NBER_ID}.pdf"
+        STATUS=$?
+        if (( $STATUS == 0 ))
+        then
+            printf "Download succeeded. Now let me sleep for ${SLEEP} seconds... \xF0\x9F\x98\xB4 \n"
+        else
+            printf "Download failed. Now let me sleep for ${SLEEP} seconds... \xF0\x9F\x98\xB4 \n"
+        fi
+        sleep $SLEEP
+        pdftotext -layout "${DIRECTORY}/${NBER_ID}.pdf" "${DIRECTORY}/${NBER_ID}.txt"
+        rm -rf "${DIRECTORY}/${NBER_ID}.pdf"
+    fi
     START=$(($START + 1))
 done
